requirements.txt
joblib==1.2.0
scikit-learn==1.2.2
beautifulsoup4 == 4.12.2
bs4 == 0.0.1
dnspython == 2.3.0
PySocks == 1.7.1
requests == 2.28.2
